fix: access token not refreshed when using stream (#240)
* fix stream access token expired

* fix async access token
